Extract common methods to trait
Extracted common methods for introspecting event manager listeners to a new
trait, EventManagerIntrospectionTrait, and updated all test cases that consumed
one or more of them to use the trait instead.
